When updating build pack bits, delay blobstore deletions to allow current stagings to succeed.
